Relative Path Support for Icons
Changed the behavior of icons to use relative paths as this is what the
spec calls for.

The path will be built with the baseUrl from the manifest to show the
preview image and send the API request.

Added a Helper to show the icon which will clean up the base url if need
be so we don't have slashes all over the place.