Added DMC resources and changed Mozilla->MDN
Mozilla was misspelled in this article (Mozzilla), so I changed it, but then realized that MDN is a far more well-known term to refer to the docs than Mozilla. So I changed it to MDN.

I also realized there was no mention of our own docs here, so I added a section.