[Validations] Categories fix (#3889)
fixes #[issue number]
https://github.com/pwa-builder/PWABuilder/issues/3866

## PR Type
<!-- Please uncomment one ore more that apply to this PR -->

Bugfix
<!-- - Feature -->
<!-- - Code style update (formatting) -->
<!-- - Refactoring (no functional changes, no api changes) -->
<!-- - Build or CI related changes -->
<!-- - Documentation content changes -->
<!-- - Sample app changes -->
<!-- - Other... Please describe: -->


## Describe the current behavior?
<!-- Please describe the current behavior that is being modified or link
to a relevant issue. -->
We were testing for A) that categories is an array and B) that it was a
"standard category"

## Describe the new behavior?
We no longer test that each category is a "standard" category because:
A) The criteria for the list of categories on the W3C spec is unclear,
not used by any browser, and not used by any store.
B) Start moving towards each test only testing one thing, of which
categories is one of the last

## PR Checklist

- [x ] Test: run `npm run test` and ensure that all tests pass
- [x ] Target main branch (or an appropriate release branch if
appropriate for a bug fix)
- [ x] Ensure that your contribution follows [standard accessibility
guidelines](https://docs.microsoft.com/en-us/microsoft-edge/accessibility/design).
Use tools like https://webhint.io/ to validate your changes.


## Additional Information

---------

Co-authored-by: Justin Willis (HE / HIM) <juwillis@microsoft.com>