Giving developers control over the Windows Icon Background color option in the packaging form (#3892)
fixes #3861
<!-- Link to relevant issue (for ex: "fixes #1234") which will
automatically close the issue once the PR is merged -->

## PR Type
Bugfix / Feature
<!-- - Code style update (formatting) -->
<!-- - Refactoring (no functional changes, no api changes) -->
<!-- - Build or CI related changes -->
<!-- - Documentation content changes -->
<!-- - Sample app changes -->
<!-- - Other... Please describe: -->


## Describe the current behavior?
<!-- Please describe the current behavior that is being modified or link
to a relevant issue. -->
Right now, we use the developers `manifest.background_color` as a
default and `"transparent"` as a fallback for the Windows icon
background color.

## Describe the new behavior?
Now we have a color pick in the windows packaging form that gives the
developer complete control over that field.

## PR Checklist
- [x] Test: run `npm run test` and ensure that all tests pass
- [x] Target main branch (or an appropriate release branch if
appropriate for a bug fix)
- [x] Ensure that your contribution follows [standard accessibility
guidelines](https://docs.microsoft.com/en-us/microsoft-edge/accessibility/design).
Use tools like https://webhint.io/ to validate your changes.


## Additional Information
This PR contains some other minor bug fixes and updates to the way we
use the `sl-select` component as it was rewritten by shoelace recently.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Justin Willis <jgw9617@gmail.com>
Co-authored-by: Justin Willis (HE / HIM) <juwillis@microsoft.com>
Co-authored-by: Zach Teutsch <88554871+zateutsch@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Nikola Metulev <nmetulev@users.noreply.github.com>
Co-authored-by: Amrutha Srinivasan <amrutha.srinivasan95@gmail.com>
Co-authored-by: Judah Gabriel Himango <judahgabriel@gmail.com>