We no longer wait for the service worker to enable packaging button  (#4012)
fixes #4011
<!-- Link to relevant issue (for ex: "fixes #1234") which will
automatically close the issue once the PR is merged -->

## PR Type
Feature

## Describe the current behavior?
<!-- Please describe the current behavior that is being modified or link
to a relevant issue. -->
We wait on the SW test to finish before enabling the package for stores
button even though it has no impact on the decision to enable that
button anymore.

## Describe the new behavior?
As soon as the manifest and security tests are deemed valid enough to
package, the button enables.

## PR Checklist
- [x] Test: run `npm run test` and ensure that all tests pass
- [x] Target main branch (or an appropriate release branch if
appropriate for a bug fix)
- [x] Ensure that your contribution follows [standard accessibility
guidelines](https://docs.microsoft.com/en-us/microsoft-edge/accessibility/design).
Use tools like https://webhint.io/ to validate your changes.

## Additional Information

Co-authored-by: Nikola Metulev <nmetulev@users.noreply.github.com>