Edge Side Panel update and other bug fixes (#4049)
- Edge side panel receives the new tooltip in the manifest editor and
its priority is swapped to recommended instead of optional.
- @amrutha95, When you click package on the windows form for the first
time, the value of resoureLanguage is converted from a list to a string.
So hen you would click package back to back in the windows form, the
code was calling `.join()` on a string and it was crashing. Now we check
if its a string still (from the first time) or if they went back and
selected more languages it will be a list again and will convert it to a
string for packaging.
- The "Edit in manifest" button wasn't working as expected because some
code got deleted. I just put that code back to make it work as expected.