Sub-domains bug fixed (#4146)
Added a validation that allows PWAs to be loaded from folders also.
Currently, your PWA must be located in the main folder, if not, you will
face the issue that I reported in this GitHub issue:
https://github.com/pwa-builder/PWABuilder/issues/4085

This patch fixes that.

Here is an example of the compiled and fixed lib:
https://fanmixco.github.io/toastmasters-timer/js/lib/pwaupdate.min.js

This is a case where the new update lib is working:
https://fanmixco.github.io/toastmasters-timer

fixes #4085
<!-- Link to relevant issue (for ex: "fixes #1234") which will
automatically close the issue once the PR is merged -->

## PR Type

Bugfix

<!-- Please uncomment one ore more that apply to this PR -->

<!-- - Bugfix -->
<!-- - Feature -->
<!-- - Code style update (formatting) -->
<!-- - Refactoring (no functional changes, no api changes) -->
<!-- - Build or CI related changes -->
<!-- - Documentation content changes -->
<!-- - Sample app changes -->
<!-- - Other... Please describe: -->


## Describe the current behavior?
<!-- Please describe the current behavior that is being modified or link
to a relevant issue. -->

It considers the main domain only, sub-domains are ignored.

## Describe the new behavior?

It registers the web worker and recognize the default URL regardless if
it's the main folder (empty domain) or a subfolder.

## PR Checklist

- [x] Test: run `npm run test` and ensure that all tests pass
- [x] Target main branch (or an appropriate release branch if
appropriate for a bug fix)
- [x] Ensure that your contribution follows [standard accessibility
guidelines](https://docs.microsoft.com/en-us/microsoft-edge/accessibility/design).
Use tools like https://webhint.io/ to validate your changes.


## Additional Information