[CLI] Change Usage Tracking Behavior (#4379)
@jgw96 This contains the changes we talked about. I also added a readme
for display on NPM.

If you could just update to **0.0.15** and make sure none of the
commands hang that would be great.

Summary of changes:

The bulk of each command and the analytics process are now running on
separate processes. I also detached the analytics process from the
parent, and if I understand/tested correctly, we should get analytics
regardless of whether or not the command was cancelled, unless they
manage to cancel before we make the analytics call.