fix(): remove un-needed brittle test (#4457)
fixes #4449 
<!-- Link to relevant issue (for ex: "fixes #1234") which will
automatically close the issue once the PR is merged -->

## PR Type
<!-- Please uncomment one ore more that apply to this PR -->

Bugfix
<!-- - Feature -->
<!-- - Code style update (formatting) -->
<!-- - Refactoring (no functional changes, no api changes) -->
<!-- - Build or CI related changes -->
<!-- - Documentation content changes -->
<!-- - Sample app changes -->
<!-- - Other... Please describe: -->


## Describe the current behavior?
<!-- Please describe the current behavior that is being modified or link
to a relevant issue. -->
We had a test that tried to match the specific scores for the demo app,
however this has already broken multiple times not because of bugs, but
because we changed our scoring logic a bit. We should not have tests
that we have to update every time we update our scoring logic, and
ensuring the demo app passes is already checked with other tests we
have.

## Describe the new behavior?
No longer testing for specific scores for the demo app.

## PR Checklist

- [ x] Test: run `npm run test` and ensure that all tests pass
- [ x] Target main branch (or an appropriate release branch if
appropriate for a bug fix)
- [ x] Ensure that your contribution follows [standard accessibility
guidelines](https://docs.microsoft.com/en-us/microsoft-edge/accessibility/design).
Use tools like https://webhint.io/ to validate your changes.


## Additional Information

Co-authored-by: Justin Willis (HE / HIM) <juwillis@microsoft.com>
Co-authored-by: Jaylyn Barbee <51131738+Jaylyn-Barbee@users.noreply.github.com>