#4172/4610/4612: Manifest Editor bug fixes (#4615)
fixes #4172
fixes #4610 
fixes #4612

Fixed a few bugs with how errors are displayed in the manifest editor.

I had to change the manifest tests a little to get this to work as
expected, so @jgw96 or @Jaylyn-Barbee one of y'all should probably take
a close look at this PR to make sure I didn't cause anything unexpected.

---------

Co-authored-by: Justin Willis <jgw9617@gmail.com>