fix(): temp fix for packaging as we work through backend issues (#4852)
fixes #[issue number] 
<!-- Link to relevant issue (for ex: "fixes #1234") which will
automatically close the issue once the PR is merged -->

## PR Type
<!-- Please uncomment one ore more that apply to this PR -->

 Bugfix 


## Describe the current behavior?
<!-- Please describe the current behavior that is being modified or link
to a relevant issue. -->
Due to an issue in our backend, the packaging process has some issues
currently, which we need to fix our deployments to fix. Because of this,
the quicker fix is to get a temp front-end fix in.

## Describe the new behavior
This will not enable widgets if the app does not actually implement
widgets, BUT, by setting this flag, we avoid the broken code on our
backend.

## PR Checklist

- [x ] Test: run `npm run test` and ensure that all tests pass
- [ x] Target main branch (or an appropriate release branch if
appropriate for a bug fix)
- [ x] Ensure that your contribution follows [standard accessibility
guidelines](https://docs.microsoft.com/en-us/microsoft-edge/accessibility/design).
Use tools like https://webhint.io/ to validate your changes.


## Additional Information