fix service worker registration timeout #4553 (#4923)
fixes #4553 
<!-- Link to relevant issue (for ex: "fixes #1234") which will
automatically close the issue once the PR is merged -->

## PR Type
<!-- Please uncomment one ore more that apply to this PR -->

Bugfix
<!-- - Feature -->
<!-- - Code style update (formatting) -->
<!-- - Refactoring (no functional changes, no api changes) -->
<!-- - Build or CI related changes -->
<!-- - Documentation content changes -->
<!-- - Sample app changes -->
<!-- - Other... Please describe: -->


## Describe the current behavior?
<!-- Please describe the current behavior that is being modified or link
to a relevant issue. -->
When a page has the service worker resource but does not register it
properly on the page, the `AuditServiceWorker` request works properly
but the page's `Report` request returns that there is no service worker,
this generates a race condition and mismatch where it first appears that
the page has a service worker and then (with the second request) updates
the status to show that no service worker was found.

## Describe the new behavior?
A warning banner notifies that a mismatch has been found, possibly
pointing to an incorrect registration of the service worker.

## PR Checklist

- [ ] Test: run `npm run test` and ensure that all tests pass
- [ ] Target main branch (or an appropriate release branch if
appropriate for a bug fix)
- [ ] Ensure that your contribution follows [standard accessibility
guidelines](https://docs.microsoft.com/en-us/microsoft-edge/accessibility/design).
Use tools like https://webhint.io/ to validate your changes.


## Additional Information