fix unexpected close dialog when a tooltip is open and escape key is pressed #4262 (#4937)
fixes #4262 
<!-- Link to relevant issue (for ex: "fixes #1234") which will
automatically close the issue once the PR is merged -->

## PR Type
<!-- Please uncomment one ore more that apply to this PR -->

Bugfix
<!-- - Feature -->
<!-- - Code style update (formatting) -->
<!-- - Refactoring (no functional changes, no api changes) -->
<!-- - Build or CI related changes -->
<!-- - Documentation content changes -->
<!-- - Sample app changes -->
<!-- - Other... Please describe: -->


## Describe the current behavior?
<!-- Please describe the current behavior that is being modified or link
to a relevant issue. -->
When the manifest modal is opened, it has icons that open a tooltip to
give more detail about the field to be filled. If escape is pressed when
a tooltip is open, it closes the modal along with the tooltip. When the
intention would be that only the tooltip is closed.

## Describe the new behavior?
When there is a tooltip open in the manifest dialog, and the escape key
is pressed, it just closes the tooltip, and in case there is no tooltip
open, it closes the dialog as expected.

## PR Checklist

- [ ] Test: run `npm run test` and ensure that all tests pass
- [ ] Target main branch (or an appropriate release branch if
appropriate for a bug fix)
- [ ] Ensure that your contribution follows [standard accessibility
guidelines](https://docs.microsoft.com/en-us/microsoft-edge/accessibility/design).
Use tools like https://webhint.io/ to validate your changes.


## Additional Information