Fixes bug where model state wasn't updated due to input event not firing (#5038)
Fixes an issue where if you set custom values in packaging form, e.g.
you set PackageID to something non-default in the Android Form, then
closed the form and opened it again, the values in the UI would be set
to your new package ID, but the underlying model wouldn't be set to the
new value. Thus, the actual Android package you generated would be set
to the original value.

This PR fixes that bug.

## PR Type
Bugfix